requirements.txt
cryptography~=2.8
Django~=2.2.4
djangorestframework==3.10.3
markdown
django-filter
django-cors-headers~=3.1.1
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0
firebase-admin
pyrebase
psycopg2
django-rest-swagger
phonenumbers
django-phonenumber-field
drf-firebase-auth

